refactor: factorize
