+++ internal/oidc/client_credentials_test.go
@@ -14,7 +14,7 @@ import (
 	"testing"
 	"time"
 
-	jwt "github.com/golang-jwt/jwt/v5"
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
 	"github.com/ory/fosite"
@@ -23,7 +23,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"github.com/valyala/fasthttp"
-	jose "gopkg.in/square/go-jose.v2"
+	"gopkg.in/square/go-jose.v2"
 
 	"github.com/authelia/authelia/v4/internal/authorization"
 	"github.com/authelia/authelia/v4/internal/configuration/schema"
@@ -32,14 +32,6 @@ import (
 	"github.com/authelia/authelia/v4/internal/oidc"
 	"github.com/authelia/authelia/v4/internal/utils"
 )
-
-func Test(t *testing.T) {
-	types := fosite.ResponseModeTypes{fosite.ResponseModeDefault, fosite.ResponseModeQuery, fosite.ResponseModeFragment, fosite.ResponseModeQuery}
-
-	assert.True(t, types.Has(fosite.ResponseModeDefault))
-	assert.True(t, types.Has(fosite.ResponseModeQuery))
-	assert.True(t, types.Has(fosite.ResponseModeFragment))
-}
 
 func TestShouldNotRaiseErrorOnEqualPasswordsPlainText(t *testing.T) {
 	hasher, err := oidc.NewHasher()
