+++ internal/oidc/util.go
@@ -334,14 +334,17 @@ func toTime(v any, def time.Time) (t time.Time) {
 	}
 }
 
-func handleWriteAuthorizeErrorJSON(ctx context.Context, config fosite.Configurator, rw http.ResponseWriter, rfcerr *fosite.RFC6749Error) {
+func handleWriteAuthorizeErrorJSON(ctx context.Context, config fosite.Configurator, rw http.ResponseWriter, rfc *fosite.RFC6749Error) {
 	rw.Header().Set(fasthttp.HeaderContentType, headerContentTypeApplicationJSON)
 
-	js, err := json.Marshal(rfcerr)
-	if err != nil {
+	var (
+		data []byte
+		err  error
+	)
+
+	if data, err = json.Marshal(rfc); err != nil {
 		if config.GetSendDebugMessagesToClients(ctx) {
-			errorMessage := fosite.EscapeJSONString(err.Error())
-			http.Error(rw, fmt.Sprintf(`{"error":"server_error","error_description":"%s"}`, errorMessage), http.StatusInternalServerError)
+			http.Error(rw, fmt.Sprintf(`{"error":"server_error","error_description":"%s"}`, fosite.EscapeJSONString(err.Error())), http.StatusInternalServerError)
 		} else {
 			http.Error(rw, `{"error":"server_error"}`, http.StatusInternalServerError)
 		}
@@ -349,6 +352,6 @@ func handleWriteAuthorizeErrorJSON(ctx context.Context, config fosite.Configurat
 		return
 	}
 
-	rw.WriteHeader(rfcerr.CodeField)
-	_, _ = rw.Write(js)
+	rw.WriteHeader(rfc.CodeField)
+	_, _ = rw.Write(data)
 }
